pypsga.py
# pypsga.py

# Import necessary packages
import os
import json
from git import Repo
from config import BASE_DIR, TEMPLATES_DIR, DEFAULT_PROJECT_STRUCTURE, DEFAULT_TEMPLATE, VCS, AI_OPTIMIZATION_PARAMS, DOC_AUTOMATION_PARAMS
from jinja2 import Environment, FileSystemLoader

# Load templates
env = Environment(loader=FileSystemLoader(TEMPLATES_DIR))

def create_directory_structure(structure, root_dir):
    """
    Function to create directory structure
    """
    for key, value in structure.items():
        dir_path = os.path.join(root_dir, key)
        os.makedirs(dir_path, exist_ok=True)
        if isinstance(value, dict):
            create_directory_structure(value, dir_path)

def generate_project_structure(template=DEFAULT_TEMPLATE):
    """
    Function to generate project structure
    """
    template = env.get_template(f'{template}.json')
    project_structure = json.loads(template.render())
    create_directory_structure(project_structure, BASE_DIR)

def analyze_project_structure():
    """
    Function to analyze project structure
    """
    # TODO: Implement project structure analysis

def integrate_with_vcs():
    """
    Function to integrate with version control system
    """
    if VCS == 'git':
        Repo.init(BASE_DIR)

def optimize_project_structure():
    """
    Function to optimize project structure using AI
    """
    # TODO: Implement AI-driven project structure optimization

def adapt_project_structure():
    """
    Function to adapt project structure based on ongoing changes
    """
    # TODO: Implement dynamic project structure adaptation

def automate_documentation():
    """
    Function to automate documentation
    """
    # TODO: Implement documentation automation

def main():
    """
    Main function
    """
    generate_project_structure()
    analyze_project_structure()
    integrate_with_vcs()
    optimize_project_structure()
    adapt_project_structure()
    automate_documentation()

if __name__ == "__main__":
    main()
